Crate substrate_primitives

Shareable Substrate types.

Re-exports

pub use hashing::blake2_128;

pub use hashing::blake2_256;

pub use hashing::twox_64;

pub use hashing::twox_128;

pub use hashing::twox_256;

pub use self::hash::convert_hash;

pub use crypto::DeriveJunction;

pub use crypto::Pair;

Modules

crypto
Cryptographic utilities.

ed25519
Simple Ed25519 API.

hash
A fixed hash type.

hashing
Hashing functions.

hexdisplay
Wrapper type for byte collections that outputs hex.

offchain
Offchain workers types

sandbox
Definition of a sandbox environment.

sr25519
Simple sr25519 (Schnorr-Ristretto) API.

storage
Contract execution data.

u32_trait
An u32 trait with "values" as impl'd types.

uint
An unsigned fixed-size integer.

Macros

map
Initialize a key-value collection from array.

Structs

Blake2Hasher
Concrete implementation of Hasher using Blake2b 256-bit hashes

Bytes
Hex-serialized shim for Vec<u8>.

ChangesTrieConfiguration
Substrate changes trie configuration.

H160
Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256
Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512
Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

OpaqueMetadata
Stores the encoded RuntimeMetadata for the native side as opaque type.

U256
Little-endian large integer type

Enums

ExecutionContext
Context for executing a call into the runtime.

NativeOrEncoded
Something that is either a native or an encoded value.

NeverNativeValue
A value that is never in a native representation. This is type is useful in conjuction with NativeOrEncoded.

Traits

Hasher
Trait describing an object that can hash a slice of bytes. Used to abstract other types over the hashing algorithm. Defines a single hash method and an Out associated type with the necessary bounds.