Trait substrate_primitives::crypto::IsWrappedBy

pub trait IsWrappedBy<Outer>: From<Outer> + Into<Outer> {
    fn from_ref(outer: &Outer) -> &Self;

    fn from_mut(outer: &mut Outer) -> &mut Self;
}

One type is wrapped by another.

Required methods

fn from_ref(outer: &Outer) -> &Self

Get a reference to the inner from the outer.

fn from_mut(outer: &mut Outer) -> &mut Self

Get a mutable reference to the inner from the outer.

Loading content...