Trait substrate_primitives::crypto::UncheckedFrom

pub trait UncheckedFrom<T> {
    fn unchecked_from(t: T) -> Self;
}

Similar to From, except that the onus is on the part of the caller to ensure that data passed in makes sense. Basically, you're not guaranteed to get anything sensible out.

Required methods

fn unchecked_from(t: T) -> Self

Convert from an instance of T to Self. This is not guaranteed to be whatever counts as a valid instance of T and it's up to the caller to ensure that it makes sense.

Loading content...