Trait substrate_primitives::offchain::Externalities

pub trait Externalities {
    fn submit_transaction(&mut self, extrinsic: Vec<u8>) -> Result<(), ()>;

    fn new_crypto_key(&mut self, crypto: CryptoKind) -> Result<CryptoKeyId, ()>;

    fn encrypt(
        &mut self, 
        key: Option<CryptoKeyId>, 
        data: &[u8]
    ) -> Result<Vec<u8>, ()>;

    fn decrypt(
        &mut self, 
        key: Option<CryptoKeyId>, 
        data: &[u8]
    ) -> Result<Vec<u8>, ()>;

    fn sign(
        &mut self, 
        key: Option<CryptoKeyId>, 
        data: &[u8]
    ) -> Result<Vec<u8>, ()>;

    fn verify(
        &mut self, 
        key: Option<CryptoKeyId>, 
        msg: &[u8], 
        signature: &[u8]
    ) -> Result<bool, ()>;

    fn timestamp(&mut self) -> Timestamp;

    fn sleep_until(&mut self, deadline: Timestamp);

    fn random_seed(&mut self) -> [u8; 32];

    fn local_storage_set(&mut self, key: &[u8], value: &[u8]);

    fn local_storage_compare_and_set(
        &mut self, 
        key: &[u8], 
        old_value: &[u8], 
        new_value: &[u8]
    );

    fn local_storage_get(&mut self, key: &[u8]) -> Option<Vec<u8>>;

    fn http_request_start(
        &mut self, 
        method: &str, 
        uri: &str, 
        meta: &[u8]
    ) -> Result<HttpRequestId, ()>;

    fn http_request_add_header(
        &mut self, 
        request_id: HttpRequestId, 
        name: &str, 
        value: &str
    ) -> Result<(), ()>;

    fn http_request_write_body(
        &mut self, 
        request_id: HttpRequestId, 
        chunk: &[u8], 
        deadline: Option<Timestamp>
    ) -> Result<(), HttpError>;

    fn http_response_wait(
        &mut self, 
        ids: &[HttpRequestId], 
        deadline: Option<Timestamp>
    ) -> Vec<HttpRequestStatus>;

    fn http_response_headers(
        &mut self, 
        request_id: HttpRequestId
    ) -> Vec<(Vec<u8>, Vec<u8>)>;

    fn http_response_read_body(
        &mut self, 
        request_id: HttpRequestId, 
        buffer: &mut [u8], 
        deadline: Option<Timestamp>
    ) -> Result<usize, HttpError>;
}

An extended externalities for offchain workers.

Required methods

fn submit_transaction(&mut self, extrinsic: Vec<u8>) -> Result<(), ()>

Submit transaction.

The transaction will end up in the pool and be propagated to others.

fn new_crypto_key(&mut self, crypto: CryptoKind) -> Result<CryptoKeyId, ()>

Create new key(pair) for signing/encryption/decryption.

Returns an error if given crypto kind is not supported.

fn encrypt(
&mut self,
key: Option<CryptoKeyId>,
data: &[u8]
) -> Result<Vec<u8>, ()>

Encrypt a piece of data using given crypto key.

If key is None, it will attempt to use current authority key.

Returns an error if key is not available or does not exist.

fn decrypt(
&mut self,
key: Option<CryptoKeyId>,
data: &[u8]
) -> Result<Vec<u8>, ()>

Decrypt a piece of data using given crypto key.

If key is None, it will attempt to use current authority key.

Returns an error if data cannot be decrypted or the key is not available or does not exist.

fn sign(&mut self, key: Option<CryptoKeyId>, data: &[u8]) -> Result<Vec<u8>, ()>

Sign a piece of data using given crypto key.

If key is None, it will attempt to use current authority key.

Returns an error if key is not available or does not exist.

fn verify(
&mut self,
key: Option<CryptoKeyId>,
msg: &[u8],
signature: &[u8]
) -> Result<bool, ()>

Verifies that signature for msg matches given key.

Returns an Ok with true in case it does, false in case it doesn't. Returns an error in case the key is not available or does not exist or the parameters lengths are incorrect.

fn timestamp(&mut self) -> Timestamp

Returns current UNIX timestamp (in millis)

fn sleep_until(&mut self, deadline: Timestamp)

Pause the execution until deadline is reached.

fn random_seed(&mut self) -> [u8; 32]

Returns a random seed.

This is a trully random non deterministic seed generated by host environment. Obviously fine in the off-chain worker context.

fn local_storage_set(&mut self, key: &[u8], value: &[u8])

Sets a value in the local storage.

Note this storage is not part of the consensus, it's only accessible by offchain worker tasks running on the same machine. It IS persisted between runs.

fn local_storage_compare_and_set(
&mut self,
key: &[u8],
old_value: &[u8],
new_value: &[u8]
)

Sets a value in the local storage if it matches current value.

Since multiple offchain workers may be running concurrently, to prevent data races use CAS to coordinate between them.

Note this storage is not part of the consensus, it's only accessible by offchain worker tasks running on the same machine. It IS persisted between runs.

fn local_storage_get(&mut self, key: &[u8]) -> Option<Vec<u8>>

Gets a value from the local storage.

If the value does not exist in the storage None will be returned. Note this storage is not part of the consensus, it's only accessible by offchain worker tasks running on the same machine. It IS persisted between runs.

fn http_request_start(
&mut self,
method: &str,
uri: &str,
meta: &[u8]
) -> Result<HttpRequestId, ()>

Initiaties a http request given HTTP verb and the URL.

Meta is a future-reserved field containing additional, parity-codec encoded parameters. Returns the id of newly started request.

fn http_request_add_header(
&mut self,
request_id: HttpRequestId,
name: &str,
value: &str
) -> Result<(), ()>

Append header to the request.

fn http_request_write_body(
&mut self,
request_id: HttpRequestId,
chunk: &[u8],
deadline: Option<Timestamp>
) -> Result<(), HttpError>

Write a chunk of request body.

Writing an empty chunks finalises the request. Passing None as deadline blocks forever.

Returns an error in case deadline is reached or the chunk couldn't be written.

fn http_response_wait(
&mut self,
ids: &[HttpRequestId],
deadline: Option<Timestamp>
) -> Vec<HttpRequestStatus>

Block and wait for the responses for given requests.

Returns a vector of request statuses (the len is the same as ids). Note that if deadline is not provided the method will block indefinitely, otherwise unready responses will produce DeadlineReached status.

Passing None as deadline blocks forever.

fn http_response_headers(
&mut self,
request_id: HttpRequestId
) -> Vec<(Vec<u8>, Vec<u8>)>

Read all response headers.

Returns a vector of pairs (HeaderKey, HeaderValue).

fn http_response_read_body(
&mut self,
request_id: HttpRequestId,
buffer: &mut [u8],
deadline: Option<Timestamp>
) -> Result<usize, HttpError>

Read a chunk of body response to given buffer.

Returns the number of bytes written or an error in case a deadline is reached or server closed the connection. Passing None as a deadline blocks forever.

Loading content...