Struct substrate_primitives::sr25519::Pair

pub struct Pair(_);

An Schnorrkel/Ristretto x25519 ("sr25519") key pair.

Methods

impl Pair

pub fn from_entropy(entropy: &[u8], password: Option<&str>) -> (Pair, [u8; 32])

Make a new key pair from binary data derived from a valid seed phrase.

This uses a key derivation function to convert the entropy into a seed, then returns the pair generated from it.