Struct substrate_primitives::sr25519::Public

pub struct Public(pub [u8; 32]);

An Schnorrkel/Ristretto x25519 ("sr25519") public key.

Methods

impl Public

pub fn from_raw(data: [u8; 32]) -> Self

A new instance from the given 32-byte data.

NOTE: No checking goes on to ensure this is a real public key. Only use it if you are certain that the array actually is a pubkey. GIGO!

pub fn from_slice(data: &[u8]) -> Self

A new instance from the given slice that should be 32 bytes long.

NOTE: No checking goes on to ensure this is a real public key. Only use it if you are certain that the array actually is a pubkey. GIGO!

pub fn from_h256(x: H256) -> Self

A new instance from an H256.

NOTE: No checking goes on to ensure this is a real public key. Only use it if you are certain that the array actually is a pubkey. GIGO!

pub fn into_raw_vec(self) -> Vec<u8>

Return a Vec<u8> filled with raw data.

pub fn as_slice(&self) -> &[u8]

Return a slice filled with raw data.

pub fn as_array_ref(&self) -> &[u8; 32]

Return a slice filled with raw data.