Module substrate_primitives::crypto

Cryptographic utilities.

Enums

DeriveJunction
A since derivation junction description. It is the single parameter used when creating a new secret key from an existing secret key and, in the case of SoftRaw and SoftIndex a new public key from an existing public key.

Infallible
The infallible type.

PublicError
An error type for SS58 decoding.

SecretStringError
An error with the interpretation of a secret.

Constants

DEV_ADDRESS
The address of the associated root phrase for our publicly known keys.

DEV_PHRASE
The root phrase for our publicly known keys.

JUNCTION_ID_LEN
The length of the junction identifier. Note that this is also referred to as the CHAIN_CODE_LENGTH in the context of Schnorrkel.

Traits

Derive
Derivable key trait.

Pair
Trait suitable for typical cryptographic PKI key pair type.

Ss58Codec
Key that can be encoded to/from SS58.

UncheckedFrom
Similar to From, except that the onus is on the part of the caller to ensure that data passed in makes sense. Basically, you're not guaranteed to get anything sensible out.

UncheckedInto
The counterpart to UncheckedFrom.